LPS-165844 Move interactions filter to topper as we don't need to use it with topper empty items
Link to StackOverflow issue: https://stackoverflow.com/questions/49318497/google-chrome-simultaneously-smooth-scrollintoview-with-more-elements-doesn
